finish integrating automation 2.0

+++ contracts/src/interfaces/IKeeperRegistry.sol
@@ -5,5 +5,5 @@ import {AutomationRegistryInterface} from
     "@chainlink/contracts/src/v0.8/automation/interfaces/2_0/AutomationRegistryInterface2_0.sol";
 
 interface IKeeperRegistry is AutomationRegistryInterface {
-    function getMaxPaymentForGas(uint256 gasLimit) external view returns (uint96 maxPayment);
+    function getMaxPaymentForGas(uint32 gasLimit) external view returns (uint96 maxPayment);
 }
